Lots of conflicts seen, some unclear (why are all the updated files conflicted?)
Some confusing (FirmataDevice.cs)

# Conflicts:
#	azure-pipelines.yml
#	eng/ArduinoCsCI.cmd
#	src/System.Device.Gpio/System/Device/Gpio/Drivers/RaspberryPi3Driver.cs
#	src/devices/Arduino/ArduinoBoard.cs
#	src/devices/Arduino/ArduinoGpioControllerDriver.cs
#	src/devices/Arduino/Encoder7Bit.cs
#	src/devices/Arduino/ExtendedCommandHandler.cs
#	src/devices/Arduino/FirmataDevice.cs
#	src/devices/Arduino/ReconnectingNetworkStream.cs
#	src/devices/Arduino/samples/ApiChecker/Arduino.sample.csproj
#	src/devices/Arduino/samples/ApiChecker/TestCases.cs
#	src/devices/Arduino/samples/Monitor/Arduino.Monitor.csproj
#	src/devices/Arduino/tests/Arduino.Tests.csproj
#	src/devices/Board/I2cBusExtensions.cs
#	tools/ArduinoCsCompiler/ArduinoCsCompiler.csproj
#	tools/ArduinoCsCompiler/ArduinoCsCompiler.sln
#	tools/ArduinoCsCompiler/ArduinoImplementationAttribute.cs
#	tools/ArduinoCsCompiler/ArduinoMethodDeclaration.cs
#	tools/ArduinoCsCompiler/CompilerCommandHandler.cs
#	tools/ArduinoCsCompiler/CompilerSettings.cs
#	tools/ArduinoCsCompiler/Debugger.cs
#	tools/ArduinoCsCompiler/EquatableMethod.cs
#	tools/ArduinoCsCompiler/ErrorManager.cs
#	tools/ArduinoCsCompiler/ExecutionSet.cs
#	tools/ArduinoCsCompiler/Frontend/CompilerOptions.cs
#	tools/ArduinoCsCompiler/Frontend/CompilerRun.cs
#	tools/ArduinoCsCompiler/Frontend/Frontend.csproj
#	tools/ArduinoCsCompiler/Frontend/Program.cs
#	tools/ArduinoCsCompiler/Hal/ArduinoNativeGpioDriver.cs
#	tools/ArduinoCsCompiler/KnownTypeTokens.cs
#	tools/ArduinoCsCompiler/MethodFlags.cs
#	tools/ArduinoCsCompiler/MicroCompiler.cs
#	tools/ArduinoCsCompiler/README.md
#	tools/ArduinoCsCompiler/Runtime/MiniArray.cs
#	tools/ArduinoCsCompiler/Runtime/MiniBuffer.cs
#	tools/ArduinoCsCompiler/Runtime/MiniCultureInfo.cs
#	tools/ArduinoCsCompiler/Runtime/MiniEnum.cs
#	tools/ArduinoCsCompiler/Runtime/MiniInterop.Kernel32.cs
#	tools/ArduinoCsCompiler/Runtime/MiniMonitor.cs
#	tools/ArduinoCsCompiler/Runtime/MiniThread.cs
#	tools/ArduinoCsCompiler/VariableKind.cs
#	tools/ArduinoCsCompiler/WriteRuntimeCoreData.cs
#	tools/ArduinoCsCompiler/samples/WeatherStation/WeatherStation.csproj
#	tools/ArduinoCsCompiler/tests/ArduinoNativeLibraryTest.cs
#	tools/ArduinoCsCompiler/tests/ArduinoTestBase.cs
#	tools/ArduinoCsCompiler/tests/FirmataIlExecutorTests.cs
#	tools/ArduinoCsCompiler/tests/GarbageCollectorTests.cs
#	tools/ArduinoCsCompiler/tests/MiniExamples.cs

+++ tools/ArduinoCsCompiler/CompilerCommandHandler.cs
@@ -344,7 +344,7 @@ namespace ArduinoCsCompiler
         {
             FirmataIlCommandSequence sequence = new FirmataIlCommandSequence(ExecutorCommand.DeclareMethod);
             sequence.SendInt32(declarationToken);
-            sequence.SendUInt14((byte)methodFlags);
+            sequence.SendUInt14(methodFlags);
             sequence.WriteByte(maxStack);
             sequence.WriteByte(argCount);
             sequence.SendInt32((int)nativeMethod);
